+++ app/test/e2e/dataset-get-includes-responses/dataset-get-include-all-admin.js
@@ -24,8 +24,6 @@ const datasetGetIncludeAllAdmin = dataset => ({
         status: 'saved',
         published: true,
         overwrite: true,
-        verified: false,
-        blockchain: {},
         mainDateField: null,
         env: 'production',
         geoInfo: false,