+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v1.3.1
+
+## 19/05/2020
+
+- Use secondary mongo nodes for read operations.
+- Disable mongo unified topology.
+
 # v1.3.0
 
 ## 09/04/2020